+++ packages/angular-workspace/nimble-angular/table-column/text/tests/nimble-table-column-text.directive.spec.ts
@@ -34,6 +34,7 @@ describe('NimbleTableColumnText', () => {
                         sort-direction="${TableColumnSortDirection.ascending}"
                         sort-index="0"
                         sorting-disabled
+                        sort-by-field-name="field2"
                         group-index="0"
                         grouping-disabled
                         placeholder="Custom placeholder"
@@ -101,6 +102,11 @@ describe('NimbleTableColumnText', () => {
             expect(nativeElement.sortingDisabled).toBeTrue();
         });
 
+        it('will use template string value for sortByFieldName', () => {
+            expect(directive.sortByFieldName).toBe('field2');
+            expect(nativeElement.sortByFieldName).toBe('field2');
+        });
+
         it('will use template string values for fractionalWidth', () => {
             expect(directive.fractionalWidth).toBe(2);
             expect(nativeElement.fractionalWidth).toBe(2);
@@ -143,6 +149,7 @@ describe('NimbleTableColumnText', () => {
                         [sort-direction]="sortDirection"
                         [sort-index]="sortIndex"
                         [sorting-disabled]="sortingDisabled"
+                        [sort-by-field-name]="sortByFieldName"
                         [group-index]="groupIndex"
                         [grouping-disabled]="groupingDisabled"
                         [placeholder]="placeholder"
@@ -163,6 +170,7 @@ describe('NimbleTableColumnText', () => {
             public sortDirection: TableColumnSortDirection = TableColumnSortDirection.ascending;
             public sortIndex: number | null = 0;
             public sortingDisabled = false;
+            public sortByFieldName = 'field2';
             public groupIndex: number | null = 0;
             public groupingDisabled = false;
             public placeholder = 'Custom placeholder';
@@ -282,6 +290,17 @@ describe('NimbleTableColumnText', () => {
             expect(nativeElement.sortingDisabled).toBeTrue();
         });
 
+        it('can be configured with property binding for sortByFieldName', () => {
+            expect(directive.sortByFieldName).toBe('field2');
+            expect(nativeElement.sortByFieldName).toBe('field2');
+
+            fixture.componentInstance.sortByFieldName = 'anotherField';
+            fixture.detectChanges();
+
+            expect(directive.sortByFieldName).toBe('anotherField');
+            expect(nativeElement.sortByFieldName).toBe('anotherField');
+        });
+
         it('can be configured with property binding for fractionalWidth', () => {
             expect(directive.fractionalWidth).toBe(2);
             expect(nativeElement.fractionalWidth).toBe(2);
@@ -387,6 +406,7 @@ describe('NimbleTableColumnText', () => {
                         [attr.sort-direction]="sortDirection"
                         [attr.sort-index]="sortIndex"
                         [attr.sorting-disabled]="sortingDisabled"
+                        [attr.sort-by-field-name]="sortByFieldName"
                         [attr.group-index]="groupIndex"
                         [attr.grouping-disabled]="groupingDisabled"
                         [attr.placeholder]="placeholder"
@@ -407,6 +427,7 @@ describe('NimbleTableColumnText', () => {
             public sortDirection: TableColumnSortDirection = TableColumnSortDirection.ascending;
             public sortIndex: number | null = 0;
             public sortingDisabled = false;
+            public sortByFieldName = 'field2';
             public groupIndex: number | null = 0;
             public groupingDisabled = false;
             public placeholder = 'Custom placeholder';
@@ -524,6 +545,17 @@ describe('NimbleTableColumnText', () => {
 
             expect(directive.sortingDisabled).toBeTrue();
             expect(nativeElement.sortingDisabled).toBeTrue();
+        });
+
+        it('can be configured with attribute binding for sortByFieldName', () => {
+            expect(directive.sortByFieldName).toBe('field2');
+            expect(nativeElement.sortByFieldName).toBe('field2');
+
+            fixture.componentInstance.sortByFieldName = 'anotherField';
+            fixture.detectChanges();
+
+            expect(directive.sortByFieldName).toBe('anotherField');
+            expect(nativeElement.sortByFieldName).toBe('anotherField');
         });
 
         it('can be configured with attribute binding for fractionalWidth', () => {