+++ packages/nimble-components/src/table/types.ts
@@ -214,7 +214,7 @@ export interface TableRowState<TData extends TableRecord = TableRecord> {
     groupColumn?: TableColumn;
     isParentRow: boolean;
     isLoadingChildren: boolean;
-    slots: SlotMetadata[];
+    requestedSlots: SlotMetadata[];
     resolvedRowIndex?: number;
 }